Move ProgramOutput to dsl package
