+++ website/src/main/kotlin/org/pointyware/xyz/site/utils/ProgramInputs.kt
@@ -1,5 +1,7 @@
 package org.pointyware.xyz.site.utils
 
+import org.pointyware.xyz.site.dsl.ProgramOutput
+
 /**
  * Defines the available parameters for controlling program behavior.
  * This model exists primarily as a helper object for processing arguments progressively.
